+++ context.go
@@ -6,6 +6,7 @@ package jug
 
 import (
 	"io"
+	"net/http"
 	"time"
 )
 
@@ -58,6 +59,19 @@ type Context interface {
 	// MustBindJSONV tries to bind the request body from JSON to the given object. If that fails the request is aborted with 400.
 	// If it succeeds the provided validator function is invoked.
 	MustBindJSONV(obj any, validator func() error) bool
+
+	// Request gets the original http request
+	Request() *http.Request
+	// Writer gets the http response writer
+	Writer() http.ResponseWriter
+	//ClientIP implements one best effort algorithm to return the real client IP.
+	//It calls c.RemoteIP() under the hood, to check if the remote IP is a trusted proxy or not.
+	//If it is it will then try to parse the headers defined in Engine.RemoteIPHeaders (defaulting to [X-Forwarded-For, X-Real-Ip]).
+	//If the headers are not syntactically valid OR the remote IP does not correspond to a trusted proxy,
+	//the remote IP (coming from Request.RemoteAddr) is returned.
+	ClientIP() string
+	//RemoteIP parses the IP from Request.RemoteAddr, normalizes and returns the IP (without the port).
+	RemoteIP() string
 
 	// Status sets the response status code.
 	Status(code int) Context